add day 11и правки дня 9

+++ src/main/java/day9/Task2/Rectangle.java
@@ -10,11 +10,11 @@ public class Rectangle extends Figure {
         this.height = height;
     }
 
-    public void setWidth(int width) {
+    public void setWidth(double width) {
         this.width = width;
     }
 
-    public void setHeight(int height) {
+    public void setHeight(double height) {
         this.height = height;
     }
